+++ project/history.go
@@ -29,7 +29,20 @@ func FullVersion() string {
 // To cut a new release add a release to the front of this slice then run the
 // release tagging script: ./scripts/tag_release.sh
 var History relic.ImmutableHistory = relic.NewHistory("Hyperledger Burrow", "https://github.com/hyperledger/burrow").
-	MustDeclareReleases(
+	MustDeclareReleases("0.21.0 - 2018-08-21",
+		`### Changed
+- Upgraded to Tendermint 0.23.0
+- Validator Set Power now takes Address
+- RPC/TM config renamed to RPC/Info
+
+### Added
+- Burrow deploy creates devdoc
+- Docker image has org.label-schema labels
+
+### Fixed
+- Upgrade to IAVL 0.10.0 and load previous versions immutably on boot - for chains with a long history > 20 minute load times could be observed because every previous root was being loaded from DB rather than lightweight version references as was intended
+- Metrics server does not panic on empty block metas and recovers from other panics
+`,
 		"0.20.1 - 2018-08-17",
 		`### Changed
 - The snatives functions have new signatures; string arguments are now string, not byte32.
@@ -38,6 +51,7 @@ var History relic.ImmutableHistory = relic.NewHistory("Hyperledger Burrow", "htt
 ### Fixed
 - TxExecutions that were exceptions (for example those that were REVERTed) will no longer have their events emitted from ExecutionEventsServer.GetEvents. They remain stored in state for the time being.
 - CallTxSim and CallCodeSim now take same code path as real transactions (via CallContext)
+- Release our mempool signing lock once transactions have been CheckTx'd' to massively increase throughput.
 
 ### Added
 - Upgraded to Tendermint [0.22.8](https://github.com/tendermint/tendermint/compare/v0.22.4...v0.22.8) (from 0.22.4).
@@ -47,10 +61,6 @@ var History relic.ImmutableHistory = relic.NewHistory("Hyperledger Burrow", "htt
 - Ability to set --external-address on burrow configure and burrow start
 - Ability to set various command line options on burrow configure and burrow start and by BURROW_ prefixed environment variables
 - Exposed Tendermint SeedMode option
-
-
-### Fixed
-- Release our mempool signing lock once transactions have been CheckTx'd' to massively increase throughput.
 `,
 		"0.20.0 - 2018-07-24",
 		`This is a major (pre-1.0.0) release that introduces the ability to change the validator set through GovTx, transaction execution history, and fuller GRPC endpoint.