+++ CHANGELOG.md
@@ -1,4 +1,19 @@
 # [Hyperledger Burrow](https://github.com/hyperledger/burrow) Changelog
+## [0.21.0] - 2018-08-21
+### Changed
+- Upgraded to Tendermint 0.23.0
+- Validator Set Power now takes Address
+- RPC/TM config renamed to RPC/Info
+
+### Added
+- Burrow deploy creates devdoc
+- Docker image has org.label-schema labels
+
+### Fixed
+- Upgrade to IAVL 0.10.0 and load previous versions immutably on boot - for chains with a long history > 20 minute load times could be observed because every previous root was being loaded from DB rather than lightweight version references as was intended
+- Metrics server does not panic on empty block metas and recovers from other panics
+
+
 ## [0.20.1] - 2018-08-17
 ### Changed
 - The snatives functions have new signatures; string arguments are now string, not byte32.
@@ -7,6 +22,7 @@
 ### Fixed
 - TxExecutions that were exceptions (for example those that were REVERTed) will no longer have their events emitted from ExecutionEventsServer.GetEvents. They remain stored in state for the time being.
 - CallTxSim and CallCodeSim now take same code path as real transactions (via CallContext)
+- Release our mempool signing lock once transactions have been CheckTx'd' to massively increase throughput.
 
 ### Added
 - Upgraded to Tendermint [0.22.8](https://github.com/tendermint/tendermint/compare/v0.22.4...v0.22.8) (from 0.22.4).
@@ -16,10 +32,6 @@
 - Ability to set --external-address on burrow configure and burrow start
 - Ability to set various command line options on burrow configure and burrow start and by BURROW_ prefixed environment variables
 - Exposed Tendermint SeedMode option
-
-
-### Fixed
-- Release our mempool signing lock once transactions have been CheckTx'd' to massively increase throughput.
 
 
 ## [0.20.0] - 2018-07-24
@@ -254,6 +266,7 @@ This release marks the start of Eris-DB as the full permissioned blockchain node
   - [Blockchain] Fix getBlocks to respect block height cap.
 
 
+[0.21.0]: https://github.com/hyperledger/burrow/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/hyperledger/burrow/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/hyperledger/burrow/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/hyperledger/burrow/compare/v0.18.1...v0.19.0